requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
openai==1.3.0
python-dotenv==1.0.0
moviepy==1.0.3
pydub==0.25.1
loguru==0.7.2
pytest==7.4.3
python-slugify==8.0.1
